backend/requirements.txt
# FastAPI & server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1
python-multipart==0.0.12

# Env & validation
pydantic-settings==2.6.1
email-validator==2.2.0

# Scraper
httpx==0.28.1
beautifulsoup4==4.12.3
